+++ src/AuthModules/ThrottledLogins/OrgStatus.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Swe\SpaceSDK\AuthModules\ThrottledLogins;
+
+use GuzzleHttp\Exception\GuzzleException;
+use Swe\SpaceSDK\AbstractApi;
+
+/**
+ * Class OrgStatus
+ *
+ * @package Swe\SpaceSDK\AuthModules\ThrottledLogins
+ * @author Luca Braun <l.braun@s-w-e.com>
+ */
+class OrgStatus extends AbstractApi
+{
+    /**
+     * Returns date and time until which the organization are throttled.
+     *
+     * @param array $response
+     * @return array
+     * @throws GuzzleException
+     */
+    public function getOrganizationThrottlingStatus(array $response = []): array
+    {
+        $uri = 'auth-modules/throttled-logins/org-status';
+
+        return $this->client->get($this->buildUrl($uri), $response);
+    }
+
+    /**
+     * Resets date and time until which the organization are throttled.
+     *
+     * @return void
+     * @throws GuzzleException
+     */
+    public function resetOrganizationThrottling(): void
+    {
+        $uri = 'auth-modules/throttled-logins/org-status';
+
+        $this->client->delete($this->buildUrl($uri));
+    }
+}